# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/users/me/account/password": {
      "put": {
        "tags": [
          "user-controller"
        ],
        "summary": "Put new password",
        "operationId": "putPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "New password is not valid or supplied old password is incorrect",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "User is not authenticated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "Password updated"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "USER"
            ]
          }
        ]
      }
    },
...
    "/api/apartments": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "summary": "Search in apartments by query",
        "operationId": "search",
        "parameters": [
          {
            "name": "searchQuery",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid query",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Apartments received",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApartmentCardDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "manager-apartment-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApartmentForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "MANAGER"
            ]
          }
        ]
      }
    },
...
    "/api/users/me/account": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get me as user account",
        "operationId": "getAccount",
        "responses": {
          "403": {
            "description": "User is not authenticated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "User received",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserAccountDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "USER"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "user-controller"
        ],
        "summary": "Delete my account",
        "operationId": "deleteAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "User is not authenticated or supplied password is incorrect",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "USER"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "user-controller"
        ],
        "summary": "Update my account",
        "operationId": "changeAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeUserForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "User is not authenticated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "User updated"
          },
          "400": {
            "description": "Passed data to update is incorrect",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "USER"
            ]
          }
        ]
      }
    },
    "/api/users/me/account/notifications": {
      "patch": {
        "tags": [
          "user-controller"
        ],
        "summary": "Update notification settings",
        "operationId": "changeNotificationSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeNotificationSettingsForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "User is not authenticated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "400": {
            "description": "Passed data to update is incorrect",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "204": {
            "description": "Notification settings updated"
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "USER"
            ]
          }
        ]
      }
    },
...
    "/api/apartments/{id}": {
      "get": {
        "tags": [
          "apartment-controller"
        ],
        "summary": "Get an apartment by id",
        "operationId": "getOne_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Apartment not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "Apartment received",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApartmentDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id format",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "manager-apartment-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "MANAGER"
            ]
          }
        ]
      },
      "patch": {
        "tags": [
          "manager-apartment-controller"
        ],
        "operationId": "change",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeApartmentForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "security": [
          {
            "bearerAuth": [
              "MANAGER"
            ]
          }
        ]
      }
    },
...
    "/api/users/me": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "summary": "Get me as user",
        "operationId": "getOne",
        "responses": {
          "403": {
            "description": "User is not authenticated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "200": {
            "description": "User received",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [
              "USER"
            ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChangePasswordDto": {
        "required": [
          "newPassword",
          "oldPassword"
        ],
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "ApartmentId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
      "PutPriceForm": {
        "required": [
          "amount",
          "person"
        ],
        "type": "object",
        "properties": {
          "person": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "maximum": 100000,
            "exclusiveMaximum": false,
            "minimum": 0,
            "exclusiveMinimum": false,
            "type": "number"
          },
          "apartmentId": {
            "$ref": "#/components/schemas/ApartmentId"
          }
        }
      },
...
      "ApartmentOptions": {
        "type": "object",
        "properties": {
          "mealsIncluded": {
            "type": "boolean"
          },
          "kitchen": {
            "type": "boolean"
          },
          "bathroom": {
            "type": "boolean"
          },
          "wifi": {
            "type": "boolean"
          },
          "refrigerator": {
            "type": "boolean"
          }
        }
      },
      "CreateApartmentForm": {
        "required": [
          "apartmentClass",
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ApartmentId"
          },
          "name": {
            "type": "string"
          },
          "apartmentClass": {
            "type": "string",
            "enum": [
              "COMFORT",
              "STANDARD"
            ]
          },
          "options": {
            "$ref": "#/components/schemas/ApartmentOptions"
          },
          "prices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PutPriceForm"
            }
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateRoomForm"
            }
          },
          "instances": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateApartmentInstanceForm"
            }
          }
        }
      },
      "CreateApartmentInstanceForm": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 2147483647,
            "minLength": 3,
            "type": "string"
          },
          "bookingICalUrl": {
            "type": "string"
          },
          "apartmentId": {
            "$ref": "#/components/schemas/ApartmentId"
          }
        }
      },
      "CreateBedForm": {
        "required": [
          "size",
          "type"
        ],
        "type": "object",
        "properties": {
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "BUNK",
              "CONNECTED",
              "TRANSFORMER"
            ]
          }
        }
      },
      "CreateRoomForm": {
        "required": [
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 2147483647,
            "minLength": 3,
            "type": "string"
          },
          "beds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateBedForm"
            }
          },
          "type": {
            "type": "string",
            "enum": [
              "BEDROOM",
              "LIVING_ROOM",
              "MEETING_ROOM"
            ]
          }
        }
      },
...
      "ChangeUserForm": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        }
      },
      "ChangeNotificationSettingsForm": {
        "type": "object",
        "properties": {
          "receiveOrderEmails": {
            "type": "boolean"
          },
          "receiveNewsEmails": {
            "type": "boolean"
          }
        }
      },
...
      "ChangeApartmentForm": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "apartmentClass": {
            "type": "string",
            "enum": [
              "COMFORT",
              "STANDARD"
            ]
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "MANAGER"
            ]
          }
        }
      },
...
      "NotificationSettings": {
        "type": "object",
        "properties": {
          "receiveOrderEmails": {
            "type": "boolean"
          },
          "receiveNewsEmails": {
            "type": "boolean"
          }
        }
      },
      "UserAccountDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "MANAGER"
            ]
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "notificationSettings": {
            "$ref": "#/components/schemas/NotificationSettings"
          }
        }
      },
...
      "SearchQuery": {
        "type": "object",
        "properties": {
          "checkIn": {
            "type": "string",
            "format": "date"
          },
          "checkOut": {
            "type": "string",
            "format": "date"
          },
          "guests": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "COMFORT",
              "STANDARD"
            ]
          }
        }
      },
      "ApartmentCardDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "photo": {
            "type": "string"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "people": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number"
          }
        }
      },
      "ApartmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "options": {
            "$ref": "#/components/schemas/ApartmentOptions"
          },
          "rating": {
            "type": "number",
            "format": "double"
          },
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomDto"
            }
          },
          "prices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceDto"
            }
          }
        }
      },
      "BedDto": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "BUNK",
              "CONNECTED",
              "TRANSFORMER"
            ]
          }
        }
      },
      "PriceDto": {
        "type": "object",
        "properties": {
          "person": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "RoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "BEDROOM",
              "LIVING_ROOM",
              "MEETING_ROOM"
            ]
          },
          "beds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BedDto"
            }
          }
        }
      },
...
      "DeleteMeDto": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}